Dynamic range is now exactly 5.0A. Updated fieldmap to include 0 stage for field adjust. Fixed DAC out of bounds check and ADC RD CH command for the recent switch to signed integers

+++ projects/rev_d_shim/software/shim-test/include/commands/command_helper.h
@@ -80,9 +80,9 @@ typedef struct command_context {
 
 // Helper function to convert Amps to signed DAC units
 // Maps -5.0A -> -32767, 0A -> 0, 5.0A -> 32767
-int16_t amps_to_dac(float amps);
+int16_t amps_to_dac(double amps);
 // Helper function to convert signed DAC units to Amps
-float dac_to_amps(int16_t dac_value);
+double dac_to_amps(int16_t dac_value);
 // Basic parsing and validation utilities
 uint32_t parse_value(const char* str, char** endptr);
 int parse_board_number(const char* str);
